reworked queries for getting game data, preparing to pull in other changes
